+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Button in the link window to copy link URL to clipboard ([#6])
 - Group storage
+- Search for the channel view ([#7])
 
 ### Fixed
 
@@ -164,3 +165,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#4]: https://gitlab.com/Schmiddiii/flare/-/issues/4
 [#5]: https://gitlab.com/Schmiddiii/flare/-/issues/5
 [#6]: https://gitlab.com/Schmiddiii/flare/-/issues/6
+[#7]: https://gitlab.com/Schmiddiii/flare/-/issues/7